Implemented SearchList component
